+++ source/gameengine/Converter/KX_LibLoadStatus.cpp
@@ -27,11 +27,9 @@
 #include "KX_LibLoadStatus.h"
 #include "PIL_time.h"
 
-KX_LibLoadStatus::KX_LibLoadStatus(class KX_BlenderSceneConverter* kx_converter,
-				class KX_KetsjiEngine* kx_engine,
+KX_LibLoadStatus::KX_LibLoadStatus(class KX_KetsjiEngine* kx_engine,
 				class KX_Scene* merge_scene,
 				const std::string& path) :
-			m_converter(kx_converter),
 			m_engine(kx_engine),
 			m_mergescene(merge_scene),
 			m_data(nullptr),
@@ -92,11 +90,6 @@ void KX_LibLoadStatus::RunProgressCallback()
 	}
 #endif
 #endif
-}
-
-class KX_BlenderSceneConverter *KX_LibLoadStatus::GetConverter()
-{
-	return m_converter;
 }
 
 class KX_KetsjiEngine *KX_LibLoadStatus::GetEngine()
