+++ source/gameengine/Converter/KX_LibLoadStatus.h
@@ -33,7 +33,6 @@ class KX_LibLoadStatus : public PyObjectPlus
 {
 	Py_Header
 private:
-	class KX_BlenderSceneConverter*	m_converter;
 	class KX_KetsjiEngine*			m_engine;
 	class KX_Scene*					m_mergescene;
 	void*							m_data;
@@ -52,16 +51,14 @@ private:
 #endif
 
 public:
-	KX_LibLoadStatus(class KX_BlenderSceneConverter* kx_converter,
-						class KX_KetsjiEngine* kx_engine,
-						class KX_Scene* merge_scene,
-						const std::string& path);
+	KX_LibLoadStatus(class KX_KetsjiEngine* kx_engine,
+					 class KX_Scene* merge_scene,
+					 const std::string& path);
 
 	void Finish(); // Called when the libload is done
 	void RunFinishCallback();
 	void RunProgressCallback();
 
-	class KX_BlenderSceneConverter *GetConverter();
 	class KX_KetsjiEngine *GetEngine();
 	class KX_Scene *GetMergeScene();
