+++ source/gameengine/Ketsji/BL_Action.cpp
@@ -38,7 +38,7 @@
 
 // These three are for getting the action from the logic manager
 #include "KX_Scene.h"
-#include "KX_BlenderSceneConverter.h"
+#include "KX_BlenderConverter.h"
 #include "SCA_LogicManager.h"
 
 extern "C" {
@@ -161,13 +161,13 @@ bool BL_Action::Play(const std::string& name,
 	ClearControllerList();
 
 	// Create an SG_Controller
-	SG_Controller *sg_contr = BL_CreateIPO(m_action, m_obj, kxscene->GetSceneConverter());
+	SG_Controller *sg_contr = BL_CreateIPO(m_action, m_obj, kxscene);
 	m_sg_contr_list.push_back(sg_contr);
 	m_obj->GetSGNode()->AddSGController(sg_contr);
 	sg_contr->SetNode(m_obj->GetSGNode());
 
 	// World
-	sg_contr = BL_CreateWorldIPO(m_action, kxscene->GetBlenderScene()->world, kxscene->GetSceneConverter());
+	sg_contr = BL_CreateWorldIPO(m_action, kxscene->GetBlenderScene()->world, kxscene);
 	if (sg_contr) {
 		m_sg_contr_list.push_back(sg_contr);
 		m_obj->GetSGNode()->AddSGController(sg_contr);
@@ -175,7 +175,7 @@ bool BL_Action::Play(const std::string& name,
 	}
 
 	// Try obcolor
-	sg_contr = BL_CreateObColorIPO(m_action, m_obj, kxscene->GetSceneConverter());
+	sg_contr = BL_CreateObColorIPO(m_action, m_obj, kxscene);
 	if (sg_contr) {
 		m_sg_contr_list.push_back(sg_contr);
 		m_obj->GetSGNode()->AddSGController(sg_contr);
@@ -189,13 +189,12 @@ bool BL_Action::Play(const std::string& name,
 			continue;
 		}
 
-		KX_BlenderSceneConverter *converter = kxscene->GetSceneConverter();
-		RAS_IPolyMaterial *polymat = converter->FindCachedPolyMaterial(kxscene, mat);
+		RAS_IPolyMaterial *polymat = KX_GetActiveEngine()->GetConverter()->FindCachedPolyMaterial(kxscene, mat);
 		if (!polymat) {
 			continue;
 		}
 
-		sg_contr = BL_CreateMaterialIpo(m_action, mat, polymat, m_obj, converter);
+		sg_contr = BL_CreateMaterialIpo(m_action, mat, polymat, m_obj, kxscene);
 		if (sg_contr) {
 			m_sg_contr_list.push_back(sg_contr);
 			m_obj->GetSGNode()->AddSGController(sg_contr);
@@ -206,14 +205,14 @@ bool BL_Action::Play(const std::string& name,
 	// Extra controllers
 	if (m_obj->GetGameObjectType() == SCA_IObject::OBJ_LIGHT)
 	{
-		sg_contr = BL_CreateLampIPO(m_action, m_obj, kxscene->GetSceneConverter());
+		sg_contr = BL_CreateLampIPO(m_action, m_obj, kxscene);
 		m_sg_contr_list.push_back(sg_contr);
 		m_obj->GetSGNode()->AddSGController(sg_contr);
 		sg_contr->SetNode(m_obj->GetSGNode());
 	}
 	else if (m_obj->GetGameObjectType() == SCA_IObject::OBJ_CAMERA)
 	{
-		sg_contr = BL_CreateCameraIPO(m_action, m_obj, kxscene->GetSceneConverter());
+		sg_contr = BL_CreateCameraIPO(m_action, m_obj, kxscene);
 		m_sg_contr_list.push_back(sg_contr);
 		m_obj->GetSGNode()->AddSGController(sg_contr);
 		sg_contr->SetNode(m_obj->GetSGNode());
