+++ source/gameengine/Launcher/LA_Launcher.h
@@ -38,7 +38,7 @@
 
 class KX_Scene;
 class KX_ISystem;
-class KX_ISceneConverter;
+class KX_BlenderConverter;
 class KX_NetworkMessageManager;
 class RAS_ICanvas;
 class DEV_EventConsumer;
@@ -76,7 +76,7 @@ protected:
 	/// The rasterizer.
 	RAS_Rasterizer *m_rasterizer;
 	/// Converts Blender data files.
-	KX_ISceneConverter *m_sceneConverter;
+	KX_BlenderConverter *m_converter;
 	/// Manage messages.
 	KX_NetworkMessageManager *m_networkMessageManager;
