+++ source/gameengine/Physics/Common/PHY_IPhysicsEnvironment.h
@@ -46,6 +46,7 @@ class RAS_MeshObject;
 struct DerivedMesh;
 class KX_GameObject;
 class KX_Scene;
+class KX_BlenderSceneConverter;
 
 struct PHY_ShapeProps;
 struct PHY_MaterialProps;
@@ -228,7 +229,8 @@ public:
 
 	virtual void MergeEnvironment(PHY_IPhysicsEnvironment *other_env) = 0;
 
-	virtual void ConvertObject(KX_GameObject *gameobj,
+	virtual void ConvertObject(KX_BlenderSceneConverter *converter,
+							   KX_GameObject *gameobj,
 	                           RAS_MeshObject *meshobj,
 	                           DerivedMesh *dm,
 	                           KX_Scene *kxscene,
