+++ source/gameengine/Physics/Dummy/DummyPhysicsEnvironment.h
@@ -127,7 +127,8 @@ public:
 		// Dummy, nothing to do here
 	}
 
-	virtual void ConvertObject(KX_GameObject *gameobj,
+	virtual void ConvertObject(KX_BlenderSceneConverter *converter,
+							   KX_GameObject *gameobj,
 	                           RAS_MeshObject *meshobj,
 	                           DerivedMesh *dm,
 	                           KX_Scene *kxscene,
